+++ R/wflow_build.R
@@ -44,9 +44,9 @@
 #'   files. Useful for site-wide changes like updating the theme, navigation
 #'   bar, or any other setting in \code{_site.yml}.
 #' @param view logical (default: \code{\link{interactive}}). View the website
-#'   after building files. If only one file is built, it is opened. If more than
-#'   one file is built, the main index page is opened. Not applicable if no
-#'   files are built or if \code{dry_run = TRUE}.
+#'   with \code{\link{wflow_view}} after building files. If only one file is
+#'   built, it is opened. If more than one file is built, the main index page is
+#'   opened. Not applicable if no files are built or if \code{dry_run = TRUE}.
 #' @param seed numeric (default: 12345). The seed to set before building each
 #'   file. Passed to \code{\link{set.seed}}.
 #' @param log_dir character (default: NULL). The directory to save log files