+++ NEWS.md
@@ -1,7 +1,9 @@
 # workflowr 0.9.0.9000
 
 * `wflow_view()` can now open a webpage in the [RStudio Viewer][]
+* Main workflowr functions now accept [file globs][glob] as input (#84)
 
+[glob]: https://en.wikipedia.org/wiki/Glob_(programming)
 [rstudio-viewer]: https://rstudio.github.io/rstudio-extensions/rstudio_viewer.html
 
 # workflowr 0.9.0