+++ NEWS.md
@@ -1,8 +1,9 @@
 # workflowr 0.9.0.9000
 
-* `wflow_view()` can now open a webpage in the [RStudio Viewer][]
+* `wflow_view()` can now open a webpage in the [RStudio Viewer][rstudio-viewer]
 * Main workflowr functions now accept [file globs][glob] as input (#84)
 * `wflow_build()` automatically removes unused figure files
+* Improve documentation of `wflow_build()` arguments
 
 [glob]: https://en.wikipedia.org/wiki/Glob_(programming)
 [rstudio-viewer]: https://rstudio.github.io/rstudio-extensions/rstudio_viewer.html